Do not send a command twice in auto retry logic
In redis-py, if a connection dies after sending a request (while reading
the response) this will result in a retry. For commands that are not
idempotent or deterministic, this could lead to bugs.

This patch overrides the redis-py execute_command() to only retry in the
event the failure occurs while sending the command.

See andymccurdy/redis-py#977